Dynamically adjust number of iterations for sortHandles
Previously, this was fixed to

(sizeof(uint32_t) * CHAR_BIT + bitsPerIter - 1) / bitsPerIter

but adjusting the number of iterations to match the maximum
element of the array gives a performance increase when the
actual number of iterations is less than what would have been
needed before the change. In the test program, this amounts to
about a 25% speedup in querying.